feat(): Add `allowDrop` function to droppable components
In some use-cases, one needs to allow dropping items is some areas only,
but these areas are dynamically defined. In such situations, `dropZones`
cannot help because they are static.

This change adds a new property to `DroppableComponent`s: `allowDrop`.
It allows defining a custom boolean function to decide whether the
currently dragged item can be dropped here or not.

For instance:

```html
<div dnd-draggable [dragData]="4">4</div>
<div dnd-draggable [dragData]="6">6</div>
<div dnd-draggable [dragData]="9">9</div>

<div dnd-droppable [allowDrop]="allowOnlyMultiplesOf(userInput)"></div>
```

```js
userInput = 2;

allowOnlyMultiplesOf(baseInteger) {
    return (dragData) => dragData % baseInteger === 0;
}
```

This commit also includes a working example that can be added to
https://github.com/akserg/ng2-systemjs-demo