Fix bug with multi-list sortable: item dropped at bottom, and the item dragged changed with first item in the list.
If we drag and drop an item to other list, the item will be appended at the bottom of the new list. Then the first item of the list will become the item that we drag. This bug could be reproduced easily in the multi-list dnd example by putting an item in the first boxer team, then dragged an item, hover through the first team, and drop into the second team.