Fix issues with dropEnabled for container that becomes empty
If a container that had element(s) became empty, any element won't be able to be dropped to the container, because dropEnabled stays false.

This commit fixes by changing dropEnabled if the data of the container becomes empty. To do that in minimal invasion to existing code, we can use Object.observe or Proxy. Because Object.observe is deprecated, and Proxy is not supported in IE, we have to change DragDropSortableService. From only containing the array (sortableData), DragDropSortableService now contains the container (csortableContainer).